requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
opencv-python>=4.10.0.84
torch>=2.3.1
torchvision>=0.18.1
facenet-pytorch>=2.6.0
numpy>=1.26.4
firebase-admin>=6.5.0
pandas>=2.2.2
python-multipart>=0.0.9
jinja2>=3.1.4
requests>=2.25.1
tqdm>=4.62.3
python-dotenv>=1.0.0
